Configure database YAML file for localhost server
This uses the rails encrypted credentials to protect the password
information for the mysql database user.

While making this work I found 2 points to be aware of...

The mysql server started using 'brew services' is automatically
restarted. This could possibly be changed somehow, but in this case I
needed to use 'ps aux | grep mysql' and 'sudo kill -9 [PID]' to force
the server to restart.

A restart was required because in the process of securing the mysql
server I forgot the root user password. After much serching and trying I
found that adding the lines below to /etc/mysql/conf.d/mysql.cnf and
restarting the server allowed me to open in a safe mode as root and
flush privileges, thereby allowing me to alter the root user password.

/etc/mysql/conf.d/mysql.cnf
```
[mysqld]
skip-grant-tables
```