Insert a Next page into the view using @page param
Note that the parameters are always strings, irrespective of the
information being provided to the view. If you need to use it as a
different type (e.g. a number) it must be coerced into the correct Type.

In this cast the `.to_i` method is used on the param value to enable
addition of two numbers.